Fix Linux + Windows CI build errors on integration (round 5)
Linux gcc/clang:
- CraftingOrderMgr.cpp: include WorldSession.h (member access into incomplete
  type 'WorldSession' at customer->GetSession()->GetAccountId()).
- Garrison.cpp: cast FOLLOWER_TYPE_GARRISON to uint8 in a second follower-type
  ternary (-Werror=extra), matching the round-4 fix.
- ChallengeModeHandler.cpp: mark the currently-unwired MythicPlusRewardBonuses
  helper [[maybe_unused]] (-Wunused-function).

Windows MSVC (/WX):
- garrison_invasion.cpp: comment the unreferenced killer / diff parameter names
  in JustDied / UpdateAI (C4100).

+++ src/server/game/Garrison/Garrison.cpp
@@ -2631,7 +2631,7 @@ void Garrison::AddFollowerXP(uint64 dbId, uint32 xp)
     WorldPackets::Garrison::GarrisonFollower oldFollowerState = follower->PacketInfo;
 
     GarrFollowerEntry const* followerEntry = sGarrFollowerStore.LookupEntry(follower->PacketInfo.GarrFollowerID);
-    uint8 followerTypeID = followerEntry ? followerEntry->GarrFollowerTypeID : FOLLOWER_TYPE_GARRISON;
+    uint8 followerTypeID = followerEntry ? followerEntry->GarrFollowerTypeID : uint8(FOLLOWER_TYPE_GARRISON);
 
     follower->PacketInfo.Xp += xp;
